Make advanced search work with picking search fields from dropdown
There are six options now, and we show 4 fields at any one time, and let users change which field we use in each.
We ensure that no field can be picked twice by eliminating it from dropdowns where it is not used.
We find all non-empty fields and construct a query from this.