graph packing algorithm: readable view of dependencies when nonlinear
We track which is the highest tier a gap has a line on. We start from shortest links, and draw them at the lowest possible tier, making sure that it is at least 1 tier higher than the highest tier it passes.
This ensures that no lines are on the same level at the same gap between PAs.