exclude ids when looking for a type key when getting document
When getting life courses the returned result is an object with a
lifecourse_id and person_appearance. The introduction of looking
dynamically for typeKeys broke the resolver for lifecourses so none of
them could be seen. It is now fixed, by not including ids.

We might need a more robust way of implementing getDocument in the
future.