Fix alignment of sizes of sorting dropdowns with size of "Alle filtre" button
The line height had been changed to 1.5 per default, which resulted in dropdowns becoming larger, because of sizing of current item in the dropdown. Fixing this to 1 realigns the sizes of the dropdowns and button. Especially visible as the height of the sort order button and sort field dropdown were misaligned