Fix logout on refresh
When refreshing the page, a login fetch is attempted, but the default callback url wasnt set, so it would just use the host - but on test and in prod this isnt an allowed callback target.
So! we need to use the baseUrl that we conveniently already get when we do a login normally. Eaz peaz