Fix error where refresh loses authentication state in angular app.
If we have an indication from localstorage that we expect to be logged in, but the SPA doesnt retrieve a user from Auth0, trigger the login flow.
In most cases this will auto-authenticate via Auth0 and end up on the same page anyway. In a few cases the login may have expired, and the user will be asked to reauthenticate.